impact calcul in file utils_eco
